+++ cyberdrop_dl/csv_logs.py
@@ -5,20 +5,24 @@ import csv
 import dataclasses
 import logging
 from collections import defaultdict
+from pathlib import Path
 from typing import TYPE_CHECKING, Any, Self
 
+from cyberdrop_dl import constants
 from cyberdrop_dl.exceptions import get_origin
 from cyberdrop_dl.logs import log_spacer
 from cyberdrop_dl.utils import json
+from cyberdrop_dl.utils.filepath import sanitize_filename
 
 if TYPE_CHECKING:
+    import datetime
     from collections.abc import Iterable, Iterator
-    from pathlib import Path
 
     from yarl import URL
 
-    from cyberdrop_dl.managers.manager import Manager
-    from cyberdrop_dl.url_objects import MediaItem
+    from cyberdrop_dl.clients.response import AbstractResponse
+    from cyberdrop_dl.manager import Manager
+    from cyberdrop_dl.url_objects import AbsoluteHttpURL, MediaItem
 
 
 logger = logging.getLogger(__name__)
@@ -27,7 +31,7 @@ _CSV_DELIMITER = ","
 
 
 @dataclasses.dataclass(slots=True, kw_only=True)
-class LogFiles:
+class CSVFiles:
     main_log: Path
     last_post_log: Path
     unsupported_urls_log: Path
@@ -43,18 +47,22 @@ class LogFiles:
 
 
 @dataclasses.dataclass(slots=True)
-class LogManager:
-    files: LogFiles
+class CSVLogsManager:
+    files: CSVFiles
     task_group: asyncio.TaskGroup = dataclasses.field(init=False, default_factory=asyncio.TaskGroup)
     _file_locks: dict[Path, asyncio.Lock] = dataclasses.field(
         init=False, default_factory=lambda: defaultdict(asyncio.Lock)
     )
     _has_headers: set[Path] = dataclasses.field(init=False, default_factory=set)
     _ready: bool = dataclasses.field(init=False, default=False)
+    _responses_folder: Path = dataclasses.field(init=False)
+
+    def __post_init__(self) -> None:
+        self._responses_folder = self.files.main_log.parent / "cdl_responses"
 
     @classmethod
     def from_manager(cls, manager: Manager) -> Self:
-        files = LogFiles(
+        files = CSVFiles(
             main_log=manager.config.settings.logs.main_log,
             last_post_log=manager.config.settings.logs.last_forum_post,
             unsupported_urls_log=manager.config.settings.logs.unsupported_urls,
@@ -129,6 +137,22 @@ class LogManager:
             self._write_to_csv(self.files.scrape_error_log, url=url, error=error_message, origin=origin)
         )
 
+    def write_response(
+        self,
+        url: AbsoluteHttpURL,
+        response: AbstractResponse[Any],
+        exc: Exception | None = None,
+    ):
+        _ = self.task_group.create_task(
+            asyncio.to_thread(
+                _write_resp_to_disk,
+                self._responses_folder,
+                url,
+                response,
+                exc,
+            )
+        )
+
     async def update_last_forum_post(self, input_file: Path) -> None:
         """Updates the last forum post."""
 
@@ -190,3 +214,27 @@ def _update_last_forum_post(input_file: Path, last_post_log: Path) -> None:
 
     with input_file.open("w", encoding="utf8") as f:
         f.write("\n".join(updated_urls))
+
+
+def _write_resp_to_disk(
+    folder: Path,
+    url: AbsoluteHttpURL,
+    response: AbstractResponse[Any],
+    exc: Exception | None = None,
+) -> None:
+    ext = ".json" if "json" in response.content_type else ".html"
+    file = _prepare_resp_file(folder, url, response.created_at, ext)
+    try:
+        _ = file.write_text(response.create_report(exc), "utf8")
+    except OSError as e:
+        logger.warning(f"Unable to write response from {url} to disk ({e!r})")
+    else:
+        logger.debug(f"Saved response from {url} to '{file}'")
+
+
+def _prepare_resp_file(folder: Path, url: AbsoluteHttpURL, created_at: datetime.datetime, ext: str = ".html") -> Path:
+    max_stem_len = 245 - len(str(folder)) + len(constants.STARTUP_TIME_STR) + 10
+    log_date = created_at.strftime(constants.LOGS_DATETIME_FORMAT)
+    path_safe_url = sanitize_filename(Path(str(url)).as_posix().replace("/", "-"))
+    filename = f"{path_safe_url[:max_stem_len]}_{log_date}{ext}"
+    return folder / filename